requirements.txt
# Core CV / Realtime
opencv-python==4.10.0.84
mediapipe==0.10.14

# API / Server / Utils
fastapi==0.115.5
uvicorn[standard]==0.32.0
websockets==12.0
numpy==1.26.4
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3

# DB
pymongo==4.9.1

# RAG / Embeddings
sentence-transformers==3.1.1
hnswlib==0.8.0     # ← FAISS 대체(Windows pip 가능)

# (Optional) Reporting
matplotlib==3.9.2
plotly==5.24.1
